Improve the way we register the OkHttpClient adapter.
